refactor: split crate internals into modules
